refactor: add ref module
