+++ virtool_cli/cli.py
@@ -1,18 +1,16 @@
 import click
 
-from virtool_cli.ref import ref
 from virtool_cli.hmm import hmm
+from virtool_cli.ref.cli import ref
 
 
 @click.group()
 def entry():
-    """
-    ex. virtool ref --help
+    """ex. virtool ref --help
 
     note: virtool hmm is has additional bioconda dependencies,
     see documentation for installation instructions
     """
-    pass
 
 
 entry.add_command(ref)
